refactor: split template_validator.py by validation concern [skip review]
Split the 2,246-line TemplateValidator class into 5 focused modules
organized by validation concern. Each file owns both its detection
logic and error formatting — they change together, so they live together.

New files:
- template_path_validation.py — Pass 5: path existence + error formatting
- template_type_validation.py — Passes 6+7: type matching + shell types
- batch_item_validation.py — Pass 8: ${item.field} validation
- validation_utils.py — Shared infrastructure (split_template_path, etc.)

template_validator.py remains as the orchestrator (~660 lines) owning
the entry point, output extraction, and simple passes.

- Dissolved TemplateValidator class (pure namespace, all @staticmethod)
- Deleted dead code: _is_valid_syntax() (defined but never called)
- Updated all 17 consumer files (4 production + 13 test)
- Updated 10 mock.patch string targets in test_planning
- Removed 6 tests for deleted dead code

Closes #118
