+++ tests/test_runtime/test_template_validator_union_types.py
@@ -10,7 +10,7 @@ Related to Issue #66: Validator should allow nested template access for typed di
 from unittest.mock import Mock
 
 from pflow.registry import Registry
-from pflow.runtime.template_validator import TemplateValidator
+from pflow.runtime.template_validator import validate_workflow_templates
 
 
 def create_mock_registry(nodes_metadata):
@@ -73,7 +73,7 @@ class TestUnionTypeValidation:
         })
 
         # Validate - should pass without errors
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
         assert len(warnings) == 0, "dict|str union should not generate warnings"
@@ -113,7 +113,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         # No errors - str allows nested access via JSON auto-parsing
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
@@ -152,7 +152,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
         assert len(warnings) == 0, "dict|object union should not generate warnings"
@@ -188,7 +188,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
         assert len(warnings) == 0, "dict|str|int union should not generate warnings"
@@ -224,7 +224,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
         # any is trusted (explicit declaration) - no warning even with str in union
@@ -261,7 +261,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
         # dict and any are both trusted - no warning
@@ -298,7 +298,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
         assert len(warnings) == 0, "Case-insensitive Dict|Str should work"
@@ -334,7 +334,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
         assert len(warnings) == 0, "Whitespace handling should work"
@@ -370,7 +370,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         # Single dict type should allow nested access without warnings
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
@@ -407,7 +407,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
         # any is an explicit declaration - node author knows what they're doing
@@ -448,7 +448,7 @@ class TestUnionTypeValidation:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         # No nested access, so no warnings even for union types
         # dict|str is compatible with str (auto-serialization)
@@ -490,7 +490,7 @@ class TestUnionTypeEdgeCases:
             },
         })
 
-        errors, _warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, _warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         # Should still work - empty string after strip is ignored
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
@@ -530,7 +530,7 @@ class TestUnionTypeEdgeCases:
             },
         })
 
-        errors, warnings = TemplateValidator.validate_workflow_templates(workflow_ir, {}, registry)
+        errors, warnings = validate_workflow_templates(workflow_ir, {}, registry)
 
         # dict|str is compatible with str (auto-serialization)
         assert len(errors) == 0, f"Expected no errors but got: {errors}"
